scripts/requirements.txt
# Python deps for scripts/generate-brand-assets.py
cairosvg
pillow